requirements.txt
asgiref==3.8.1
beautifulsoup4==4.9.3
certifi==2024.7.4
chardet==4.0.0
click==8.1.7
colorama==0.4.6
Django==5.1.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
Flask==2.0.1
greenlet==3.0.3
idna==2.10
inflection==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
packaging==24.1
PyJWT==2.9.0
pytz==2024.2
PyYAML==6.0.2
redis==3.5.3
requests==2.25.1
setuptools==75.1.0
soupsieve==2.5
SQLAlchemy==1.4.20
sqlparse==0.5.1
tzdata==2024.2
uritemplate==4.1.1
urllib3==1.26.19
Werkzeug==3.0.3
